+++ providers/cloudflare/rest.go
@@ -9,7 +9,7 @@ import (
 	"golang.org/x/net/idna"
 
 	"github.com/StackExchange/dnscontrol/v4/models"
-	"github.com/StackExchange/dnscontrol/v4/providers/cloudflare/singleredirect"
+	"github.com/StackExchange/dnscontrol/v4/providers/cloudflare/rtypes/cfsingleredirect"
 	"github.com/cloudflare/cloudflare-go"
 )
 
@@ -307,7 +307,7 @@ func (c *cloudflareProvider) getSingleRedirects(id string, domain string) ([]*mo
 		srMatcher := pr.Expression
 		srReplacement := pr.ActionParameters.FromValue.TargetURL.Expression
 		code := int(pr.ActionParameters.FromValue.StatusCode)
-		sr := singleredirect.FromAPIData(srMatcher, srReplacement, code)
+		sr := cfsingleredirect.FromAPIData(srMatcher, srReplacement, code)
 		//sr.SRRRuleList = rulelist
 		//printer.Printf("DEBUG: DESCRIPTION = %v\n", pr.Description)
 		sr.SRDisplay = pr.Description
@@ -450,7 +450,7 @@ func (c *cloudflareProvider) getPageRules(id string, domain string) ([]*models.R
 			code)
 		r.SetTarget(raw)
 
-		cr, err := singleredirect.FromUserInput(raw, code, pr.Priority)
+		cr, err := cfsingleredirect.FromUserInput(raw, code, pr.Priority)
 		if err != nil {
 			return nil, err
 		}
