+++ providers/cloudflare/rtypes/cfsingleredirect/convert_test.go
@@ -1,4 +1,4 @@
-package singleredirect
+package cfsingleredirect
 
 import (
 	"regexp"
